backend/requirements.txt
flask
requests
python-dotenv
google-generativeai
flask-cors